Move MySiteCardAndItem cards to Card sealed class, DynamicCard below Card
