+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/MySiteInfoViewHolder.kt
@@ -3,8 +3,8 @@ package org.wordpress.android.ui.mysite
 import android.view.View
 import android.view.ViewGroup
 import org.wordpress.android.databinding.MySiteInfoCardBinding
-import org.wordpress.android.ui.mysite.MySiteCardAndItem.SiteInfoCard
-import org.wordpress.android.ui.mysite.MySiteCardAndItem.SiteInfoCard.IconState
+import org.wordpress.android.ui.mysite.MySiteCardAndItem.Card.SiteInfoCard
+import org.wordpress.android.ui.mysite.MySiteCardAndItem.Card.SiteInfoCard.IconState
 import org.wordpress.android.util.image.ImageManager
 import org.wordpress.android.util.image.ImageType.BLAVATAR
 import org.wordpress.android.util.viewBinding
