+++ WordPress/src/test/java/org/wordpress/android/ui/mysite/quickactions/QuickActionsCardBuilderTest.kt
@@ -6,7 +6,7 @@ import org.junit.Before
 import org.junit.Test
 import org.wordpress.android.BaseUnitTest
 import org.wordpress.android.R
-import org.wordpress.android.ui.mysite.MySiteCardAndItem.QuickActionsCard
+import org.wordpress.android.ui.mysite.MySiteCardAndItem.Card.QuickActionsCard
 import org.wordpress.android.ui.utils.UiString.UiStringRes
 
 @InternalCoroutinesApi
